// APWeb/src/assets/model/HMR.ts
export class HMR {
  ProductID: string = "";
  ProductName: string = "";
  UPC: string = "";
  UnitQty: number | undefined;
  UnitPrice: number | undefined;
  Date!: string;
  Action: string = "";
  SellTo: string = "";
  TransactionID!: number;
  DepartmentID: string = "";
  DepartmentName: string = "";
  SupplierID!: number;
  SupplierName: string = "";
  StoreID!: number;
  ProductType: string = "";
  ProductDesc: string = "";
  UnitOfMeasure: string = "";
  Applicant: string = "";
  Location: string = "";
  
}